Create VISION.md
Added this vision document which aims to bring more clarity to the project. I wrote it to mainly add more detail to the core features I want to implement and kind of figure out where I'm going with this project. I also wanted to set the tone of the project and app vision with some taglines at the end, the last of which I think will be the final one.