Create SUCCESS_METRICS.md
Added this document to detail how we will be able to determine the success of the app. Meeting some or all of these metrics will determine whether the app needs to be scaled to a higher user-base or development team or if there is any interest in this sort of idea for an app.